Added functionality to search to filter tags.
[MEEP] I'm dynamically fetching labels from GitHub and displaying them
into the filter/tags area. As the user searches for tags, they can click
to enable or disable the filter. As the user enables a filter, it should
filter out posts/cards in the center column. As the user resizes to
tablet or mobile, the filters/tags area moves to the nav bar.

Love to get any feedback either on UI/UX or the code implementation. I'm
sure I can do much more to optimize it!